refactor(DataTableHead): extract table head to component
Refs: #1
